requirements.txt
fastapi
uvicorn
jinja2
python-multipart
passlib[bcrypt]
sqlalchemy
pyjwt
pdfplumber
openpyxl
pandas
itsdangerous
PyMuPDF
pytesseract
python-dotenv
pillow